+++ structuretimers/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "structuretimers.apps.StructureTimersConfig"
 
-__version__ = "1.2.0a2"
+__version__ = "1.2.0a3"
 __title__ = "Structure Timers"